some refactoring to improve the appearance.
Made the inner hole of the vinyl smaller and added a element to present the pin on which it sets. Changed the shape and positioning of the tonearm based on feedback. Added some more grooves. Extracted the rays of light to a separate layer that stationary above the spinning disk rather than being part of it